+++ webapp/views/views.py
@@ -159,8 +159,9 @@ class ExtractContigs(BaseViewMixin, View):
 
     template = 'chlamdb/extract_contigs.html'
     view_name = "extract_contigs"
-    _base_colname_to_header_mapping = {"_product": "Product"}
-    table_data_accessors = [
+    _base_colname_to_header_mapping = {"_product": "Product",
+                                       "accession": "Contig"}
+    loci_data_accessors = [
             "gene",
             "_product",
             "locus_tag",
@@ -170,10 +171,16 @@ class ExtractContigs(BaseViewMixin, View):
             "start",
             "end"]
 
-    def get(self, request, genome):
-        taxid = int(genome)
+    contigs_data_accessors = [
+            "accession",
+            "length",
+            "topology",
+            "plasmid"]
 
-        descr = self.db.get_genomes_description().description.to_dict()
+    def table_headers(self, columns):
+        return [self.colname_to_header(col) for col in columns]
+
+    def get_loci_tab(self, taxid):
         prot_infos = self.db.get_proteins_info(
             [taxid], search_on="taxid", as_df=True,
             to_return=["locus_tag", "product", "gene"])
@@ -190,18 +197,46 @@ class ExtractContigs(BaseViewMixin, View):
         all_infos.gene = all_infos.gene.map(format_gene)
         all_infos.locus_tag = all_infos.locus_tag.map(format_locus)
         all_infos.orthogroup = all_infos.orthogroup.map(lambda_format_og)
-
-        organism = descr[taxid]
         all_infos["_product"] = all_infos["product"]
-        result_tabs = [
-            TabularResultTab(
-                1, "Loci",
-                table_data=all_infos[self.table_data_accessors],
-                table_headers=self.table_headers,
-                table_data_accessors=self.table_data_accessors,
-                show_badge=True),
-            ]
-        context = self.get_context(organism=organism, result_tabs=result_tabs)
+        return TabularResultTab(
+            "loci", "Loci",
+            table_data=all_infos[self.loci_data_accessors],
+            table_headers=self.table_headers(self.loci_data_accessors),
+            table_data_accessors=self.loci_data_accessors,
+            show_badge=True)
+
+    def get_contigs_tab(self, taxid):
+        contigs = self.db.get_bioentry_list(
+            taxid, terms=["bioentry_id", "accession", "length"])
+        contig_qualifiers = {
+            bioentry_id: self.db.get_bioentry_qualifiers(bioentry_id)
+                                .set_index("term")["value"]
+            for bioentry_id in contigs.index}
+        contig_infos = pd.DataFrame(
+            [[bioentry_id,
+              values["topology"],
+              "Yes" if values["plasmid"] == "1" else "No"]
+             for bioentry_id, values in contig_qualifiers.items()],
+            columns=["bioentry_id", "topology", "plasmid"])
+        contigs = contigs.merge(contig_infos, left_index=True,
+                                right_on="bioentry_id")
+        return TabularResultTab(
+            "contigs", "Contigs",
+            table_data=contigs,
+            table_headers=self.table_headers(self.contigs_data_accessors),
+            table_data_accessors=self.contigs_data_accessors,
+            show_badge=True)
+
+    def get(self, request, genome):
+        taxid = int(genome)
+        descr = self.db.get_genomes_description().description.to_dict()
+        organism = descr[taxid]
+
+        loci_tab = self.get_loci_tab(taxid)
+        contigs_tab = self.get_contigs_tab(taxid)
+
+        context = self.get_context(organism=organism,
+                                   result_tabs=[loci_tab, contigs_tab])
         return render(request, self.template, context)